refactor ApiObject
