+++ src/app/models/Collections/CommentCollection.ts
@@ -1,7 +1,7 @@
 import { CollectionBase } from 'src/app/models/collections/CollectionBase';
 import { Comment } from '../classes/Comment';
 import { ApiService } from '../../services/api.service';
-import { CommentData } from 'src/app/models/api/post';
+import { CommentData } from 'src/app/models/api/PostBaseData';
 
 export class CommentCollection extends CollectionBase<number, Comment> {
