+++ src/app/models/Collections/PostCollection.ts
@@ -1,7 +1,7 @@
 import { CollectionBase } from 'src/app/models/collections/CollectionBase';
 import { Post } from 'src/app/models/classes/Post';
 import { ApiService } from '../../services/api.service';
-import { PostData } from 'src/app/models/api/post';
+import { PostData } from 'src/app/models/api/PostBaseData';
 
 export class PostCollection extends CollectionBase<number, Post> {
