+++ src/app/models/Collections/ProjectCollection.ts
@@ -1,6 +1,6 @@
 import { CollectionBase } from './CollectionBase';
 import { Project } from '../classes/Project';
-import { ProjectData } from '../api/project';
+import { ProjectData } from '../api/ProjectData';
 import { ApiService } from '../../services/api.service';
 
 export class ProjectCollection extends CollectionBase<number, Project> {
