+++ src/app/models/classes/Comment.ts
@@ -1,5 +1,5 @@
 import { ApiService } from 'src/app/services/api.service';
-import { CommentData } from '../api/post';
+import { CommentData } from '../api/PostBaseData';
 import { PostBase } from './PostBase';
 
 export class Comment extends PostBase<CommentData> {
