--- src/app/services/collections/comment-collection.ts
@@ -1,28 +1,0 @@
-import { Injectable } from '@angular/core';
-import { CollectionBase } from './node_modules/src/app/models/_Collections/CollectionBase';
-import { Comment } from '../classes/Comment';
-import { ApiService } from '../../services/api.service';
-import { CommentData } from './node_modules/src/app/models/api/post';
-
-export class CommentCollection extends CollectionBase<number, Comment> {
-
-  constructor(
-    private readonly api: ApiService,
-  ) {
-    super();
-  }
-
-  protected async buildObject(key: number) {
-    const data = await this.api.get<CommentData>(`user/${key}`).toPromise();
-    if (!data)
-      throw new Error('Fail to build project.');
-    return new Comment(this.api, data);
-  }
-
-  async create(content: string, postId: number) {
-    const data = await this.api.post<CommentData>('comment/', { content, post_id: postId }).toPromise();
-    const comment = new Comment(this.api, data);
-    this.cache.set(comment.id, comment);
-    return comment;
-  }
-}
