--- src/app/services/collections/post-collection.ts
@@ -1,29 +1,0 @@
-import { Injectable } from '@angular/core';
-import { CollectionBase } from './node_modules/src/app/models/_Collections/CollectionBase';
-import { Post } from './node_modules/src/app/models/classes/Post';
-import { ApiService } from '../../services/api.service';
-import { PostData } from './node_modules/src/app/models/api/post';
-import { dashCaseToCamelCase } from '@angular/compiler/src/util';
-
-export class PostCollection extends CollectionBase<number, Post> {
-
-  constructor(
-    private readonly api: ApiService,
-  ) {
-    super();
-  }
-
-  protected async buildObject(key: number) {
-    const data = await this.api.get<PostData>(`post/${key}`).toPromise();
-    if (!data)
-      throw new Error('Fail to build Post.');
-    return new Post(this.api, data);
-  }
-
-  async create(content: string, projectId: number) {
-    const data = await this.api.post<PostData>('post/', { content, project_id: projectId }).toPromise();
-    const post = new Post(this.api, data);
-    this.cache.set(post.id, post);
-    return post;
-  }
-}
